DIABETES/requirements.txt
joblib==1.3.2
mysql_connector_repackaged==0.3.1
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0
reportlab==4.1.0
streamlit==1.32.1
streamlit_option_menu==0.3.12
